+++ src/config.c
@@ -27,8 +27,8 @@
 // cppcheck-suppress-end missingIncludeSystem
 
 // Include project headers
-#include "../include/config.h"
-#include "../include/coolercontrol.h"
+#include "config.h"
+#include "coolercontrol.h"
 
 /**
  * @brief Global logging implementation for all modules except main.c
