+++ src/coolercontrol.h
