+++ src/monitor.c
@@ -25,9 +25,9 @@
 // cppcheck-suppress-end missingIncludeSystem
 
 // Include project headers
-#include "../include/config.h"
-#include "../include/monitor.h"
-#include "../include/coolercontrol.h"
+#include "config.h"
+#include "monitor.h"
+#include "coolercontrol.h"
 
 /**
  * @brief Extract temperature from device status history
